gpgs: disable annoying automatic logon to Google+
I believe the user knows when to login to Google+. Give power back to the user.

No, seriously! This is not about what the game wants but what the user wants. If he f***ing
cancells the log in, we don't need to push him again to log in; nor even push him in the
first place. If he wants a leaderboard or achievements, he'll know how to get there.